requirements.txt
pyqt5
pytest
sqlalchemy
factory_boy